Improve style. Minor grammar fixes - twitter
Changed "there is a couple" to "there are a couple", even though there are two objects being referred to. This is because the intent of the phrase seems to be analogous to "a few plugins" rather than a pair of plugins that will forever be fixed as a group of two. ("there is a couple" and "there are a couple" are both grammatically correct depending on the context)